requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.10.0
chromadb==0.4.18
pypdf==3.17.1
deep-translator==1.11.4
tiktoken==0.5.2
python-dotenv==1.0.0